modules/microservices/location-consumer/requirements.txt
marshmallow==3.7.1
psycopg2-binary==2.8.5
kafka-python==2.0.2
GeoAlchemy2==0.8.4
shapely==1.7.0
